Adds a "smooth subpixel" upscaler
Thanks to drowsnug for the original implementation of a subpixel
scaler. This scaler uses the core of that one, but smooths the
transition in the new pixels added during scaling.

If you think of the scaler as 4x integer scaling of the GBA screen,
and assigning each of those pixels to a subpixel, each subpixel takes
the value of 50% the pixel assigned, 25% of the pixel to its left, and
25% of the pixel to its right.

With the pixels:

BGRBGRBGRBGR
AAAABBBBCCCC
    ^--- this subpixel will be 1/4 A[G] + 1/2 B[G] + 1/4 B[G]

The first and last pixels in each batch are not blended with the
previous / next batch to keep them sharp.